refactor: extract skipped tests into dedicated file and function
Move skippedTestsShared, skippedTestsForExpressionsRouter,
skippedTestsForTraditionalCompatibleRouter, and skippedTestsForHybrid
variables to a dedicated skipped_tests_test.go file. Add a
skippedTestsForConfig() function that builds the skip list based on
router flavor and gateway type, simplifying TestGatewayConformance.
